Refactor spotify-search into it’s own serializer
