+++ README.md
@@ -8,8 +8,8 @@ Get handy information from YouTube and Vimeo videos as well as Discogs releases 
 - https://media.now.sh/vimeo/121814744
 - https://media.now.sh/discogs/1728315
 - https://media.now.sh/spotify/3S2R0EVwBSAVMd5UMgKTL0
-- [https://media.now.sh/spotify/search/Michael Jackson - Thriller](https://media.now.sh/spotify/search/Michael Jackson - Thriller)
 - https://media.now.sh/analyse/3S2R0EVwBSAVMd5UMgKTL0
+- [https://media.now.sh/spotify-search/Michael Jackson - Thriller](https://media.now.sh/spotify-search/Michael Jackson - Thriller)
 
 The data returned will (mostly) be formatted like so:
