+++ index.js
@@ -5,9 +5,14 @@ const youtube = require('./src/serializer-youtube')
 const vimeo = require('./src/serializer-vimeo')
 const discogs = require('./src/serializer-discogs')
 const spotify = require('./src/serializer-spotify')
+const spotifySearch = require('./src/serializer-spotify-search')
 const analyse = require('./src/serializer-analyse')
 
-const serializers = {youtube, vimeo, discogs, spotify, analyse}
+const serializers = {
+	youtube, vimeo, discogs, spotify,
+	'spotify-search': spotifySearch,
+	analyse
+}
 const cors = microCors({allowMethods: ['GET']})
 
 module.exports = cors(async (request, response) => {
